+++ spec/matchers/pages_validator_matcher_spec.rb
@@ -108,6 +108,7 @@ RSpec.describe ActiveStorageValidations::Matchers::PagesValidatorMatcher do
     let(:validator_value) { 5 }
 
     it_behaves_like "base comparison validator matcher only match when exact value"
+    it_behaves_like "base comparison validator matcher equal_to rejects looser comparisons"
   end
 
   describe "#allow_blank" do